Fixed bug introduced by code cleanup Added external tool launches
git-svn-id: svn://localhost:8080/caleydo_main/trunk@1423 1f7349ae-fd9f-0d40-aeb8-9798e6c0fce3

+++ org.caleydo.core/src/org/caleydo/core/manager/picking/PickingManager.java
@@ -181,7 +181,8 @@ public class PickingManager
 				5.0, 5.0, viewport, 0); // pick width and height is set to 5
 		// (i.e. picking tolerance)
 
-		float fAspectRatio = (viewport[3] - viewport[1]) / (viewport[2] - viewport[0]);
+		float fAspectRatio = (float)(viewport[3] - viewport[1])
+						/ (float)(viewport[2] - viewport[0]);
 
 		IViewFrustum viewFrustum = canvasUser.getViewFrustum();
